backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai
PyPDF2
chromadb
sqlalchemy
psycopg2-binary
python-dotenv
google-generativeai
